Version 1.1: take the palette out of the painting, and half price bands
The mark was sampled rather than guessed at. Its paper is a warm cream
around #E7DCC1, the red in the gap of the mouth is #6F2326, the teeth
carry an olive shadow near #928A70, and the inside of the mouth is a
plum-black close to #130F17. Those four values are the whole system now.

Paper is the painting from outside: cream ground, oxblood accent,
plum-black type. It replaces the blush pink the app had been borrowing
from the site, which was never in the artwork. Mouth is the same painting
from inside, and not a recolouring of Paper — cream type on plum-black,
with the red carried up to where it reads against the dark. The accent is
lifted slightly off the canvas value: at the size a chip label is set,
#6F2326 reads brown, and red is the half of it that matters.

The site meanwhile changed what a price is. It is now a number from 1 to
4 in steps of a half, drawn as four slots where a half step lights half a
sign. Five places already sit on 2.5, and the app decoded price as an Int
— so every one of those would have thrown, and one throw fails the whole
array. The app would have gone on showing its stored copy and never said
why. Decoded as a Double now, with the site's own three-state gauge:
full, half, empty, the half slot masked down its left side so the glyph
stays whole and the row of four keeps its width.

The seed validator gains the rule that would have caught this by itself,
and the snapshot is refreshed: 73 places now.

+++ TallinnTasteBuds/Models/Place.swift
@@ -13,8 +13,9 @@ struct Place: Identifiable, Hashable, Decodable {
     let address: String
     let lat: Double
     let lng: Double
-    /// Cost band 1...4, rendered as euro signs. Not a rating: the site has no ratings.
-    let price: Int
+    /// Cost band, 1 to 4 in steps of a half, rendered as up to four euro signs
+     /// where a half step lights half a sign. Not a rating: the site has none.
+    let price: Double
     /// Type ids that exist in `taxonomy.json`.
     let types: [String]
     /// The write-up, keyed by language code. The only per-place translated field.
@@ -70,7 +71,7 @@ struct Place: Identifiable, Hashable, Decodable {
         address = try c.decode(String.self, forKey: .address)
         lat = try c.decode(Double.self, forKey: .lat)
         lng = try c.decode(Double.self, forKey: .lng)
-        price = try c.decode(Int.self, forKey: .price)
+        price = try c.decode(Double.self, forKey: .price)
         types = try c.decodeIfPresent([String].self, forKey: .types) ?? []
         blurb = try c.decodeIfPresent([String: String].self, forKey: .blurb) ?? [:]
         mustOrder = try c.decodeIfPresent([String].self, forKey: .mustOrder) ?? []
